requirements.txt
fastapi==0.108.0
uvicorn==0.15.0
httpx==0.23.0
python-dotenv==1.0.0
fastapi-mail==1.4.1
redis==5.0.1
selenium==4.16.0
httptools
httpx==0.23.0
sniffio==1.3.0
urllib3>=1.26.6
chardet>=3.0.4
retrying==1.3.4